requirements.txt
affine==2.4.0
attrs==21.4.0
boto3==1.23.6
botocore==1.26.6
Django==4.0.4
django-cors-headers==3.12.0
django-storages==1.12.3
django-tailwind==3.6.0
django-uploader==0.2.1
djangorestframework==3.13.1
docker==6.1.3
factory-boy==3.2.1
Faker==13.12.0
gunicorn==20.1.0
inflection==0.5.1
iniconfig==1.1.1
jmespath==1.0.0
packaging==21.3
Pillow==9.1.1
pluggy==1.0.0
psycopg2==2.9.3
pytest==7.1.2
pytest-django==4.5.2
pytest-factoryboy==2.4.0
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2022.1
sqlparse==0.4.4
tomli==2.0.1
typing_extensions==4.8.0
tzdata==2022.1
urllib3==1.26.9
validators==0.18.2
virtualenv==20.24.5